refactor: centralize env handling
