+++ external-data.ts
@@ -1,7 +1,8 @@
 import axios from 'axios';
+import { getEnv } from './env-helper';
 
 export async function fetchBlocknativeGas() {
-  const key = process.env.BLOCKNATIVE_KEY;
+  const key = getEnv('BLOCKNATIVE_KEY');
   if (!key) throw new Error('BLOCKNATIVE_KEY not set');
   const resp = await axios.get('https://api.blocknative.com/gasprices/blockprices', {
     headers: { Authorization: key },
@@ -11,7 +12,7 @@ export async function fetchBlocknativeGas() {
 }
 
 export async function fetchCryptoPanicNews() {
-  const key = process.env.CRYPTOPANIC_KEY;
+  const key = getEnv('CRYPTOPANIC_KEY');
   if (!key) throw new Error('CRYPTOPANIC_KEY not set');
   const url = `https://cryptopanic.com/api/v1/posts/?auth_token=${key}&kind=news`;
   const resp = await axios.get(url, { timeout: 10000 });
@@ -19,7 +20,7 @@ export async function fetchCryptoPanicNews() {
 }
 
 export async function fetchDappRadarTop() {
-  const key = process.env.DAPPRADAR_KEY;
+  const key = getEnv('DAPPRADAR_KEY');
   if (!key) throw new Error('DAPPRADAR_KEY not set');
   const resp = await axios.get('https://api.dappradar.com/4/dapps', {
     headers: { 'X-BLOBR-KEY': key },
@@ -29,7 +30,7 @@ export async function fetchDappRadarTop() {
 }
 
 export async function fetchCoinglassOI() {
-  const key = process.env.COINGLASS_KEY;
+  const key = getEnv('COINGLASS_KEY');
   if (!key) throw new Error('COINGLASS_KEY not set');
   const resp = await axios.get('https://open-api.coinglass.com/api/pro/v1/futures/openInterest', {
     headers: { coinglassSecret: key },
@@ -39,7 +40,7 @@ export async function fetchCoinglassOI() {
 }
 
 export async function queryBitquery(query: string, variables: any = {}) {
-  const key = process.env.BITQUERY_KEY;
+  const key = getEnv('BITQUERY_KEY');
   if (!key) throw new Error('BITQUERY_KEY not set');
   const resp = await axios.post(
     'https://streaming.bitquery.io/graphql',
